+++ src/main/java/org/opensearch/knn/index/codec/nativeindex/NativeIndexWriter.java
@@ -12,7 +12,6 @@ import org.apache.lucene.util.quantization.QuantizedByteVectorValues;
 import org.apache.lucene.index.FieldInfo;
 import org.apache.lucene.index.MergePolicy;
 import org.apache.lucene.index.SegmentWriteState;
-import org.apache.lucene.store.IndexOutput;
 import org.opensearch.common.Nullable;
 import org.opensearch.common.xcontent.XContentHelper;
 import org.opensearch.core.common.bytes.BytesArray;
@@ -171,8 +170,10 @@ public class NativeIndexWriter {
             fieldInfo.name,
             knnEngine.getExtension()
         );
-        try (IndexOutput output = state.directory.createOutput(engineFileName, state.context)) {
-            final IndexOutputWithBuffer indexOutputWithBuffer = new IndexOutputWithBuffer(output);
+        final IndexOutputWithBuffer indexOutputWithBuffer = new IndexOutputWithBuffer(
+            state.directory.createOutput(engineFileName, state.context)
+        );
+        try {
             final BuildIndexParams nativeIndexParams = indexParams(
                 fieldInfo,
                 indexOutputWithBuffer,
@@ -187,7 +188,9 @@ public class NativeIndexWriter {
                 knnVectorValuesSupplier.get()
             );
             indexBuilder.buildAndWriteIndex(nativeIndexParams);
-            CodecUtil.writeFooter(output);
+            CodecUtil.writeFooter(indexOutputWithBuffer.getIndexOutput());
+        } finally {
+            indexOutputWithBuffer.getIndexOutput().close();
         }
     }